% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_locate.R
\name{geo_locate}
\alias{geo_locate}
\title{Get Chinese Geolocation Data}
\usage{
geo_locate(cn_strings)
}
\arguments{
\item{cn_strings}{char vector, Chinese locations.}
}
\value{
data frame
}
\description{
Get Chinese geolocation info as a data frame. Takes as input a char vector
of Chinese location strings, returns geolocation and geocode data. For each 
input string the following variables are created: province, city, county, 
province_code, city_code, county_code.
}
\details{
Function is designed to be used with Chinese addresses and
 business names.
}
\examples{
geo_locate(c("中百超市有限公司长堤街二分店",
             "浙江省杭州市余杭区径山镇小古城村",
             "成都高梁红餐饮管理有限公司"))
}
