% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locations.R
\name{get_locations}
\alias{get_locations}
\title{Get Chinese Geolocation Codes}
\usage{
get_locations(cn_str)
}
\arguments{
\item{cn_str}{string, Chinese location}
}
\value{
Vector of geolocation data.
}
\description{
Get Chinese Geolocation Codes
}
\details{
Function is designed to be used with Chinese addresses and
 business names.
}
\examples{
get_locations("浙江省杭州市余杭区径山镇小古城村")
}
