% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_locate.R
\name{get_package_data}
\alias{get_package_data}
\title{Get geolocChina Package Data}
\usage{
get_package_data()
}
\value{
data frame
}
\description{
Returns the package data used for geolocation as a tidy data frame. This 
package uses Chinese geolocation data from this GitHub repo:
\url{https://github.com/modood/Administrative-divisions-of-China}
}
\examples{
pkg_data <- get_package_data()
}
